pyproject.toml
[project]
name = "exportio-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "dotenv>=0.9.9",
    "requests>=2.32.0",
    "djangorestframework>=3.15.0",
]
